refactor: restructure project to industry standards (layered architecture)
