+++ backend/app/schemas/chat.py
@@ -1,0 +1,16 @@
+from pydantic import BaseModel
+from datetime import datetime
+
+class MessageBase(BaseModel):
+    sender: str
+    content: str
+
+class MessageCreate(MessageBase):
+    pass
+
+class Message(MessageBase):
+    id: int
+    timestamp: datetime
+
+    class Config:
+        from_attributes = True
