+++ frontend/sirooo_ai/lib/main.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-import 'models/chat_message.dart';
-import 'services/api_service.dart';
+import 'screens/chat_screen.dart';
 
 void main() {
   runApp(const SiroooAiApp());
@@ -13,124 +12,12 @@ class SiroooAiApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'SIROOO AI',
+      debugShowCheckedModeBanner: false,
       theme: ThemeData(
         colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
         useMaterial3: true,
       ),
       home: const ChatScreen(),
-    );
-  }
-}
-
-class ChatScreen extends StatefulWidget {
-  const ChatScreen({super.key});
-
-  @override
-  State<ChatScreen> createState() => _ChatScreenState();
-}
-
-class _ChatScreenState extends State<ChatScreen> {
-  final ApiService _apiService = ApiService();
-  final TextEditingController _controller = TextEditingController();
-  List<ChatMessage> _messages = [];
-  bool _isLoading = false;
-
-  @override
-  void initState() {
-    super.initState();
-    _fetchChatHistory();
-  }
-
-  Future<void> _fetchChatHistory() async {
-    setState(() => _isLoading = true);
-    try {
-      final history = await _apiService.getChatHistory();
-      setState(() => _messages = history);
-    } catch (e) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(content: Text('Error loading history: $e')),
-      );
-    } finally {
-      setState(() => _isLoading = false);
-    }
-  }
-
-  Future<void> _handleSendMessage() async {
-    if (_controller.text.isEmpty) return;
-
-    String content = _controller.text;
-    _controller.clear();
-
-    try {
-      await _apiService.sendMessage(content);
-      _fetchChatHistory(); // Refresh history to see user message and AI response
-    } catch (e) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(content: Text('Error sending message: $e')),
-      );
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('SIROOO AI Chat'),
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-      ),
-      body: Column(
-        children: [
-          Expanded(
-            child: _isLoading 
-              ? const Center(child: CircularProgressIndicator())
-              : ListView.builder(
-                  reverse: false,
-                  itemCount: _messages.length,
-                  itemBuilder: (context, index) {
-                    final msg = _messages[index];
-                    bool isUser = msg.sender == 'user';
-                    return ListTile(
-                      title: Align(
-                        alignment: isUser ? Alignment.centerRight : Alignment.centerLeft,
-                        child: Container(
-                          padding: const EdgeInsets.all(10),
-                          decoration: BoxDecoration(
-                            color: isUser ? Colors.deepPurple[100] : Colors.grey[200],
-                            borderRadius: BorderRadius.circular(10),
-                          ),
-                          child: Text(msg.content),
-                        ),
-                      ),
-                      subtitle: Align(
-                        alignment: isUser ? Alignment.centerRight : Alignment.centerLeft,
-                        child: Text(msg.sender),
-                      ),
-                    );
-                  },
-                ),
-          ),
-          Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: Row(
-              children: [
-                Expanded(
-                  child: TextField(
-                    controller: _controller,
-                    decoration: const InputDecoration(
-                      hintText: 'Type a message...',
-                      border: OutlineInputBorder(),
-                    ),
-                  ),
-                ),
-                IconButton(
-                  icon: const Icon(Icons.send),
-                  onPressed: _handleSendMessage,
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
     );
   }
 }
