+++ frontend/sirooo_ai/lib/services/api_service.dart
@@ -1,10 +1,10 @@
 import 'dart:convert';
 import 'package:http/http.dart' as http;
 import '../models/chat_message.dart';
+import '../utils/constants.dart';
 
 class ApiService {
-  // Use 10.0.2.2 for Android Emulator, or localhost for others
-  static const String baseUrl = 'http://127.0.0.1:8000/api';
+  final String baseUrl = AppConstants.baseUrl;
 
   Future<List<ChatMessage>> getChatHistory() async {
     final response = await http.get(Uri.parse('$baseUrl/chat'));
